Fixing circular import in no_shape_geometry.py
