+++ nexus_constructor/component/chopper_shape.py
@@ -2,7 +2,8 @@ import logging
 
 from nexus_constructor.component.component_shape import ComponentShape
 from nexus_constructor.geometry.cylindrical_geometry import CylindricalGeometry
-from nexus_constructor.geometry import OFFGeometry, NoShapeGeometry
+from nexus_constructor.geometry import OFFGeometry
+from nexus_constructor.model.geometry import NoShapeGeometry
 from nexus_constructor.geometry.disk_chopper.disk_chopper_checker import (
     NexusDefinedChopperChecker,
 )
