+++ tests/test_component_factory.py
@@ -6,9 +6,9 @@ from nexus_constructor.component.pixel_shape import PixelShape
 from nexus_constructor.component.component_factory import create_component
 from nexus_constructor.geometry import (
     OFFGeometryNoNexus,
-    NoShapeGeometry,
     OFFGeometryNexus,
 )
+from nexus_constructor.model.geometry import NoShapeGeometry
 
 """
 Tests here document the conditions under which the factory creates components of different types
