+++ tests/test_no_shape_geometry.py
@@ -1,5 +1,4 @@
-from nexus_constructor.geometry import NoShapeGeometry
-from nexus_constructor.geometry.no_shape_geometry import OFFCube
+from nexus_constructor.model.geometry import OFFCube, NoShapeGeometry
 
 
 def test_GIVEN_nothing_WHEN_constructing_NoShapeGeometry_THEN_off_geometry_returns_an_OFFCube():
